// src/app/core/models/employee/employee.model.ts
export class EmployeeModel {
    birthday: string;
    companyEmail: string;
    effort: number;
    employeeCode: string;
    firstName: string;
    fullName: string;
    fullSalary: number;
    id: number;
    lastName: string;
    numberOfMonthStackExperience: number;
    phone: string;
    role: string;
    status: string;
    working: boolean;

    public constructor() {

    }
}